feat: タグ操作にJWT認証を追加

+++ resources/tag.py
@@ -1,6 +1,7 @@
 from flask.views import MethodView
 from flask_smorest import Blueprint, abort
 from sqlalchemy.exc import SQLAlchemyError
+from flask_jwt_extended import jwt_required, get_jwt_identity
 from db import db
 from models import TagModel, TodoModel
 from schema import TagSchema, PlainTagSchema
@@ -10,14 +11,18 @@ blp = Blueprint("Tags", "tags", description="Operation on tags", url_prefix="/ap
 
 @blp.route("/tags")
 class Tags(MethodView):
+    @jwt_required()
     @blp.response(200, PlainTagSchema(many=True))
     def get(self):
-        return TagModel.query.all()
+        access_user = int(get_jwt_identity())
+        return TagModel.query.filter(TagModel.user_id == access_user).all()
 
+    @jwt_required()
     @blp.arguments(TagSchema)
     @blp.response(204)
     def post(self, tag_data):
-        tag = TagModel(**tag_data)
+        access_user = int(get_jwt_identity())
+        tag = TagModel(**tag_data, user_id=access_user)
         try:
             db.session.add(tag)
             db.session.commit()
@@ -29,21 +34,29 @@ class Tags(MethodView):
 
 @blp.route("/tags/<int:tag_id>")
 class Tag(MethodView):
+    @jwt_required()
     @blp.response(200, TagSchema)
     def get(self, tag_id):
+        access_user = int(get_jwt_identity())
         tag = TagModel.query.get_or_404(tag_id)
+        if access_user != tag.user_id:
+            abort(403, message="Invalid credentials")
         return tag
 
+    @jwt_required()
     @blp.response(204)
     def delete(self, tag_id):
+        access_user = int(get_jwt_identity())
         tag = TagModel.query.get_or_404(tag_id)
+        if access_user != tag.user_id:
+            abort(403, message="Invalid credetials")
         if not tag.todos:
             try:
                 db.session.delete(tag)
                 db.session.commit()
             except SQLAlchemyError:
                 db.session.rollback()
-                abort(500, message="an error occurred while deteling the tag")
+                abort(500, message="An error occurred while deteling the tag")
             return ""
         abort(
             400,
@@ -53,18 +66,26 @@ class Tag(MethodView):
 
 @blp.route("/todos/<int:todo_id>/tag")
 class TagsInTodo(MethodView):
+    @jwt_required()
     @blp.response(200, PlainTagSchema(many=True))
     def get(self, todo_id):
+        access_user = int(get_jwt_identity())
         todo = TodoModel.query.get_or_404(todo_id)
+        if access_user != todo.user_id:
+            abort(403, message="Invalid credentials")
         return todo.tags
 
 
 @blp.route("/todos/<int:todo_id>/tag/<int:tag_id>")
 class LinkTagsToItem(MethodView):
+    @jwt_required()
     @blp.response(204)
     def post(self, todo_id, tag_id):
+        access_user = int(get_jwt_identity())
         todo = TodoModel.query.get_or_404(todo_id)
         tag = TagModel.query.get_or_404(tag_id)
+        if access_user != tag.user_id or access_user != todo.user_id:
+            abort(403, message="Invalid credentials")
         if tag not in todo.tags:
             todo.tags.append(tag)
             try:
@@ -75,10 +96,14 @@ class LinkTagsToItem(MethodView):
                 abort(500, message="An error occurred while inserting the tag.")
         return ""
 
+    @jwt_required()
     @blp.response(204)
     def delete(self, todo_id, tag_id):
+        access_user = int(get_jwt_identity())
         todo = TodoModel.query.get_or_404(todo_id)
         tag = TagModel.query.get_or_404(tag_id)
+        if access_user != tag.user_id or access_user != todo.user_id:
+            abort(403, message="Invalid credentials")
         if tag in todo.tags:
             todo.tags.remove(tag)
             try:
